+++ src/Notifier/NotifierServiceProvider.php
@@ -53,7 +53,7 @@ class NotifierServiceProvider extends ServiceProvider
      */
     public function boot()
     {
-        $path = realpath(__DIR__.'/../../');
+        $path = realpath(__DIR__.'/../');
 
         $this->package('orchestra/notifier', 'orchestra/notifier', $path);
     }